requirements.txt
aiogram==2.11
aiohttp==3.6.3
aiosqlite==0.16.0
APScheduler==3.6.3
async-timeout==3.0.1
attrs==20.3.0
Babel==2.8.0
beautifulsoup4==4.9.3
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
ctx==0.1.2
git+https://github.com/Rapptz/discord-ext-menus
discord.py==1.5.1
idna==2.10
multidict==4.7.6
pycparser==2.20
PyNaCl==1.3.0
pyTelegramBotAPI==3.7.3
pytz==2020.4
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.25.11
wavelink==0.9.6
yarl==1.5.1
